fix: isolate remaining node-sepecific path helpers
