+++ src/bindings.ts
@@ -1,18 +1,18 @@
 export let blst: Blst;
 
+// Thanks https://github.com/iliakan/detect-node/blob/master/index.esm.js
+const isNode = Object.prototype.toString.call(typeof process !== "undefined" ? process : 0) === "[object process]";
+
 // Import & assign the respective binding module.
-switch (process.env["BINDING"]) {
-  case "emscripten":
-    // TODO: factor out path
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    blst = require("../prebuild/emscripten/blst");
-    break;
-  case "swig":
-  default:
-    // eslint-disable-next-line no-case-declarations,@typescript-eslint/no-require-imports,@typescript-eslint/no-var-requires
-    const {getBinaryPath} = require("./scripts/paths");
-    // eslint-disable-next-line @typescript-eslint/no-require-imports,@typescript-eslint/no-unsafe-call
-    blst = require(getBinaryPath());
+if (!isNode || process.env.BINDING === "emscripten") {
+  // TODO: factor out path
+  // eslint-disable-next-line @typescript-eslint/no-require-imports
+  blst = require("../prebuild/emscripten/blst");
+} else {
+  // eslint-disable-next-line no-case-declarations,@typescript-eslint/no-require-imports,@typescript-eslint/no-var-requires
+  const {getBinaryPath} = require("./scripts/paths_node");
+  // eslint-disable-next-line @typescript-eslint/no-require-imports,@typescript-eslint/no-unsafe-call
+  blst = require(getBinaryPath());
 }
 
 interface MaybeAsyncModule {
