+++ src/scripts/buildBindings.ts
@@ -4,7 +4,7 @@ import {exec} from "./exec";
 import {testBindings} from "./testBindings";
 import {assertSupportedSwigVersion} from "./swig";
 import {BINDINGS_DIR, BLST_WRAP_CPP_PREBUILD} from "./paths";
-import {ensureDirFromFilepath, findBindingsFile} from "./paths_fs";
+import {ensureDirFromFilepath, findBindingsFile} from "./paths_node";
 
 /* eslint-disable no-console */
