+++ src/scripts/paths_node.ts
@@ -57,3 +57,29 @@ export function findBindingsFile(dirpath: string): string {
 
   throw Error(`Could not find bindings file. Tried:\n${filepaths.join("\n")}`);
 }
+
+/**
+ * Get binary name.
+ * name: {platform}-{arch}-{v8 version}.node
+ */
+export function getBinaryName(): string {
+  const platform = process.platform;
+  const arch = process.arch;
+  const nodeV8CppApiVersion = process.versions.modules;
+  if (!process) throw new NotNodeJsError("global object");
+  if (!platform) throw new NotNodeJsError("process.platform");
+  if (!arch) throw new NotNodeJsError("process.arch");
+  if (!process.versions.modules) throw new NotNodeJsError("process.versions.modules");
+
+  return [platform, arch, nodeV8CppApiVersion, "binding.node"].join("-");
+}
+
+export function getBinaryPath(): string {
+  return path.join(PREBUILD_DIR, getBinaryName());
+}
+
+export class NotNodeJsError extends Error {
+  constructor(missingItem: string) {
+    super(`BLST bindings loader should only run in a NodeJS context: ${missingItem}`);
+  }
+}
